Move gMongoServerConfig.getUserData and MongoServerConfig.getUserDataIfExists to the ServerConfig
