+++ loritta-discord/src/main/java/com/mrpowergamerbr/loritta/commands/vanilla/administration/AdminUtils.kt
@@ -88,7 +88,7 @@ object AdminUtils {
 			)
 		}
 
-		if (!context.legacyConfig.getUserData(context.userHandle.idLong).quickPunishment) {
+		if (!context.config.getUserData(context.userHandle.idLong).quickPunishment) {
 			replies += LoriReply(
 					context.locale["${LOCALE_PREFIX}.skipConfirmationTip", "`${context.config.commandPrefix}quickpunishment`"],
 					mentionUser = false
@@ -157,7 +157,7 @@ object AdminUtils {
 		val pipedReason = reason.split("|")
 
 		var usingPipedArgs = false
-		var skipConfirmation = context.legacyConfig.getUserData(context.userHandle.idLong).quickPunishment
+		var skipConfirmation = context.config.getUserData(context.userHandle.idLong).quickPunishment
 		var delDays = 0
 
 		var silent = false
