+++ loritta-discord/src/main/java/com/mrpowergamerbr/loritta/commands/vanilla/economy/PagarCommand.kt
@@ -45,7 +45,7 @@ class PagarCommand : AbstractCommand("pay", listOf("pagar"), CommandCategory.ECO
 			var economySource = "global"
 			var currentIdx = 0
 
-			val payerProfile = context.legacyConfig.getUserData(context.userHandle.idLong)
+			val payerProfile = context.config.getUserData(context.userHandle.idLong)
 
 			val economyConfig = transaction(Databases.loritta) {
 				loritta.getOrCreateServerConfig(context.guild.idLong).economyConfig
@@ -234,7 +234,7 @@ class PagarCommand : AbstractCommand("pay", listOf("pagar"), CommandCategory.ECO
 
 				message.addReaction("✅").queue()
 			} else {
-				val receiverProfile = context.legacyConfig.getUserData(user.idLong)
+				val receiverProfile = context.config.getUserData(user.idLong)
 
 				val beforeGiver = payerProfile.money
 				val beforeReceiver = receiverProfile.money
